+++ tidy3d/plugins/dispersion/__init__.py
@@ -1,4 +1,4 @@
 """ Imports from dispersion fitter plugin. """
 
-from .fit import DispersionFitter, StableDispersionFitter
-from .web import AdvancedFitterParam
+from .fit import DispersionFitter
+from .web import AdvancedFitterParam, StableDispersionFitter